Implementanção de Exceptions

+++ project/src/main/java/com/springProject/project/service/ProdutoService.java
@@ -1,7 +1,9 @@
 package com.springProject.project.service;
 
+import com.springProject.project.exception.RecursoNaoEncontradoException;
 import com.springProject.project.model.Produto;
 import com.springProject.project.repository.ProdutoRepository;
+import org.antlr.v4.runtime.RecognitionException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -20,14 +22,19 @@ public class ProdutoService {
         return produtoRepository.findAll();
     }
 
-    public Optional<Produto> buscarProdutoPorId(Long id) {
-        return produtoRepository.findById(id);
+    public Produto buscarProdutoPorId(Long id) {
+        return produtoRepository.findById(id)
+                .orElseThrow(() -> new RecursoNaoEncontradoException("Produto não encontrado com id: " + id));
+
     }
 
     public Produto salvarProduto(Produto produto) {
         return produtoRepository.save(produto);
     }
     public void deletarProduto(Long id) {
+        if (!produtoRepository.existsById(id)) {
+            throw new RecursoNaoEncontradoException("Produto não encontrado com id: " + id);
+        }
         produtoRepository.deleteById(id);
     }
 }
